smart_task_planner_project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
openai==1.3.7
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
